+++ docs/CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Improve
 
+* New tool `kime-check` for diagnostic kime [#270](https://github.com/Riey/kime/issues/270)
 * Correct invalid sebolsik-390 builtin layout [#261](https://github.com/Riey/kime/issues/261)
 * Fix indicator can't load icon data [#260](https://github.com/Riey/kime/issues/260)
 * Fix choseong compose bug [#263](https://github.com/Riey/kime/issues/263)